+++ trigger/conf/global_settings.py
@@ -156,6 +156,11 @@ DEFAULT_TYPES = {
 # this type.
 FALLBACK_TYPE = 'ROUTER'
 
+
+# When a manufacturer/vendor is not explicitly defined, fallback to to this
+# value.
+FALLBACK_MANUFACTURER = 'UNKNOWN'
+
 #===============================
 # Twister
 #===============================